Move transform/transform_affine from Layer to standalone functions in _transform_utils
The API is not yet stable enough to expose on the Layer class. Both
functions are now module-level in _transform_utils.py, taking
input_layer as their first parameter instead of self.
